Moved PersonalityController to the parent directory
The PersonalityController class has been relocated from the `mekhq.campaign.personnel.randomEvents.personality` package to the `mekhq.campaign.personnel.randomEvents` package. This involved updating all references in the codebase to point to the new location. Also introduced CSS tags to prevent overspill in the AtB Personnel Market.
