+++ MekHQ/src/mekhq/campaign/Campaign.java
@@ -91,7 +91,7 @@ import mekhq.campaign.personnel.marriage.AbstractMarriage;
 import mekhq.campaign.personnel.marriage.DisabledRandomMarriage;
 import mekhq.campaign.personnel.procreation.AbstractProcreation;
 import mekhq.campaign.personnel.procreation.DisabledRandomProcreation;
-import mekhq.campaign.personnel.randomEvents.personality.PersonalityController;
+import mekhq.campaign.personnel.randomEvents.PersonalityController;
 import mekhq.campaign.personnel.ranks.RankSystem;
 import mekhq.campaign.personnel.ranks.RankValidator;
 import mekhq.campaign.personnel.ranks.Ranks;
@@ -1441,6 +1441,8 @@ public class Campaign implements ITechManager {
         if (person.getAge(getLocalDate()) >= 16) {
             person.setEduHighestEducation(EducationLevel.HIGH_SCHOOL);
         }
+
+        PersonalityController.generatePersonality(person);
 
         return person;
     }
